+++ r-diagnostics.tf
@@ -2,25 +2,24 @@ module "vm_logs" {
   for_each = toset(var.use_legacy_monitoring_agent ? ["enabled"] : [])
 
   source  = "claranet/vm-logs/azurerm"
-  version = "3.0.0"
+  version = "4.1.0"
 
-  location    = var.location
-  client_name = var.client_name
   environment = var.environment
   stack       = var.stack
-
-  resource_group_name = var.resource_group_name
 
   diagnostics_storage_account_name      = var.diagnostics_storage_account_name
   diagnostics_storage_account_sas_token = var.diagnostics_storage_account_sas_token
 
-  vm_ids = [azurerm_linux_virtual_machine.vm.id]
+  vm_id = azurerm_linux_virtual_machine.vm.id
 
-  tags = {
-    environment = var.environment
-    stack       = var.stack
-    vm_name     = azurerm_linux_virtual_machine.vm.name
-  }
+  default_tags_enabled = var.default_tags_enabled
+
+  tags = merge(
+    local.default_tags,
+    {
+      vm_name = azurerm_linux_virtual_machine.vm.name
+    }
+  )
 }
 
 resource "azurerm_virtual_machine_extension" "azure_monitor_agent" {