+++ variables-tags.tf
@@ -1,0 +1,23 @@
+variable "default_tags_enabled" {
+  description = "Option to enable or disable default tags."
+  type        = bool
+  default     = true
+}
+
+variable "nic_extra_tags" {
+  description = "Extra tags to set on the network interface."
+  type        = map(string)
+  default     = {}
+}
+
+variable "public_ip_extra_tags" {
+  description = "Extra tags to set on the public IP resource."
+  type        = map(string)
+  default     = {}
+}
+
+variable "extra_tags" {
+  description = "Extra tags to set on each created resource."
+  type        = map(string)
+  default     = {}
+}